requirements.txt
influxdb
mpu6050-raspberrypi